code cleanup to package things based on our model
