+++ plugins/liquid/refined-types-plugin/src/main/kotlin/arrow/meta/plugins/liquid/phases/analysis/solver/check/model/Condition.kt
@@ -1,0 +1,9 @@
+package arrow.meta.plugins.liquid.phases.analysis.solver.check.model
+
+import org.jetbrains.kotlin.psi.KtElement
+import org.jetbrains.kotlin.psi.KtExpression
+
+/**
+ * Data type used to handle `if` and `when` without subject uniformly.
+ */
+data class Condition(val condition: KtExpression?, val body: KtExpression, val whole: KtElement)
