+++ plugins/liquid/refined-types-plugin/src/main/kotlin/arrow/meta/plugins/liquid/phases/analysis/solver/check/model/LoopPlace.kt
@@ -1,0 +1,5 @@
+package arrow.meta.plugins.liquid.phases.analysis.solver.check.model
+
+enum class LoopPlace {
+  INSIDE_LOOP, AFTER_LOOP
+}
