+++ plugins/liquid/refined-types-plugin/src/main/kotlin/arrow/meta/plugins/liquid/phases/analysis/solver/collect/model/DeclarationConstraints.kt
@@ -1,15 +1,9 @@
-package arrow.meta.plugins.liquid.phases.analysis.solver
+package arrow.meta.plugins.liquid.phases.analysis.solver.collect.model
 
 import org.jetbrains.kotlin.descriptors.DeclarationDescriptor
-import org.sosy_lab.java_smt.api.BooleanFormula
 
 data class DeclarationConstraints(
   val descriptor: DeclarationDescriptor,
   val pre: List<NamedConstraint>,
   val post: List<NamedConstraint>
-)
-
-data class NamedConstraint(
-  val msg: String,
-  val formula: BooleanFormula
 )
