+++ plugins/liquid/refined-types-plugin/src/main/kotlin/arrow/meta/plugins/liquid/phases/analysis/solver/collect/model/NamedConstraint.kt
@@ -1,0 +1,8 @@
+package arrow.meta.plugins.liquid.phases.analysis.solver.collect.model
+
+import org.sosy_lab.java_smt.api.BooleanFormula
+
+data class NamedConstraint(
+  val msg: String,
+  val formula: BooleanFormula
+)
